Language.ISOZ.Common.AbsIsoz

Description

Concrete and annotated syntax types and constructors (see clause 8 and clause 10 of ISO Z Standard)

Documentation

type NAME = String

NAME is string

data Name

Not used

Constructors

Name [Ann] String

data Specification

Specification category

Constructors

SpecSect [Ann] [Section]
sectioned specification

SpecAnony [Ann] [Paragraph]
anonymous specification

SpecEmpty [Ann]
empty specification

SpecSectA [Ann] [Section]
transformed sectioned specification

data Section

Section

Constructors

Section [Ann] String String SectionParents String [Paragraph]
Section for parsing, including Base Section and Inheriting Section . It is used in *.y and later it will be transformed to BaseSection or InheritingSection by synTraverse1Tree

BaseSection [Ann] String String String [Paragraph]
Base section without parents

InheritingSection [Ann] String String [String] String [Paragraph]
Inheriting section before syntactical transformation (synTraverse2Tree)

InheritingSectionA [Ann] NAME [NAME] [Paragraph]
Inheriting section after syntactical transformation (synTraverse2Tree)

data SectionParents

Constructors

BaseSection2 [Ann]
Used in parsing (.y) and it will be transformed to BaseSection later

InheritingSection2 [Ann] [String]
Used in parsing (.y) and it will be transformed to InheritingSection later

data Paragraph

Paragraphs

Constructors

AxPara [Ann] String [ParagraphBody] String
Paragraphs starting with ZED

AxParaA [Ann] String [Paragraph] String
Temporary use after synTraverse1Tree which transforms a list of paragraph bodies to a list of paragraphs

GivenPara [Ann] [String]
Given paragraph

GivenParaA [Ann] [NAME]
Transformed given paragraph

AxdefPara [Ann] String SchemaText String
Axiomatic definition paragraph

SchdefPara [Ann] String String SchemaText String
Schema definition paragraph

GenAxdefPara [Ann] String Formals SchemaText String
Generic axiomatic definition paragraph

GenSchdefPara [Ann] String Formals String SchemaText String
Generic schema definition paragraph

GenConjecturePara [Ann] String Formals Predicate
Generic conjecture paragraph

EPGenConjecturePara [Ann] String Formals ExprPredicate
Generic conjecture paragraph in the parsing stage where Predicate is parsed ExprPredicate

GenConjectureParaA [Ann] String [NAME] Predicate
Transformed Generic conjecture paragraph by synTraverse2Tree

HDefPara [Ann] DeclName Expression
Horizontal definition paragraph | GenHDefPara [Ann] String Formals Expression In ISO Z, it is NAME. However, it will not make ",," be parsed correctly So DeclName would be right here

GenHDefPara [Ann] DeclName Formals Expression
Generic horizontal definition paragraph

GenOpDefPara [Ann] GenName Expression
Generic operator definition paragraph

FreetypePara [Ann] [Freetype]
Free type paragraph

FreetypeParaA [Ann] [Freetype]
Transformed free type paragraph by synTraverse2Tree

ConjecturePara [Ann] String Predicate
Conjecture paragraph

EPConjecturePara [Ann] String ExprPredicate
Conjecture paragraph in the parsing stage where Predicate is parsed ExprPredicate

ConjectureParaA [Ann] String Predicate
Transformed conjecture paragraph by synTraverse2Tree

OperatorPara [Ann] OperatorTemplate
operator template paragraph

ChannelPara [Ann] [ChannelDecl]
Circus channel paragraph

ChannelFromPara [Ann] [ChannelFromDecl]
Channel from paragraph

ChannelSetPara [Ann] [ChannelSetDecl]
Channel se paragraphs annotated syntax

AxdefParaA [Ann] Expression

GenAxdefParaA [Ann] [NAME] Expression

data ParagraphBody

ParagraphBody: just used in the parsing stage to minimise more shift/reduce conflicts

Constructors

GivenPara2 [Ann] [String]

HDefPara2 [Ann] DeclName Expression

EPHDefPara2 [Ann] DeclName ExprPredicate

GenHDefPara2 [Ann] DeclName Formals Expression

EPGenHDefPara2 [Ann] DeclName Formals ExprPredicate

GenOpDefPara2 [Ann] GenName Expression

EPGenOpDefPara2 [Ann] GenName ExprPredicate

FreetypePara2 [Ann] [Freetype]

OperatorPara2 [Ann] OperatorTemplate

ChannelPara2 [Ann] [ChannelDecl]

ChannelFromPara2 [Ann] [ChannelFromDecl]

ChannelSetPara2 [Ann] [ChannelSetDecl]

ProcDeclPara2 [Ann] NAME ProcDef

GenProcDeclPara2 [Ann] [String] NAME ProcDef

data RenamePair

Rename pairf from oldname to newname

Constructors

RenamePair [Ann] DeclName DeclName

RenamePairA [Ann] NAME NAME

data ExprPredicate

A merge of Expression and Predicate. It will be transformed to Predicate and Expression shortly depending on the context

Constructors

EPForall [Ann] SchemaText ExprPredicate

EPExists [Ann] SchemaText ExprPredicate

EPExists1 [Ann] SchemaText ExprPredicate

EPEquiv [Ann] ExprPredicate ExprPredicate

EPImpl [Ann] ExprPredicate ExprPredicate

EPAnd [Ann] ExprPredicate ExprPredicate

EPOr [Ann] ExprPredicate ExprPredicate

EPNeg [Ann] ExprPredicate

EPLambda [Ann] SchemaText ExprPredicate

EPMu [Ann] SchemaText ExprPredicate

EPLocalDef [Ann] [Declaration] ExprPredicate

EPCond [Ann] ExprPredicate ExprPredicate ExprPredicate

EPComp [Ann] ExprPredicate ExprPredicate

EPPipe [Ann] ExprPredicate ExprPredicate

EPHide [Ann] ExprPredicate [DeclName]

EPProj [Ann] ExprPredicate ExprPredicate

EPPre [Ann] ExprPredicate

EPProd [Ann] [ExprPredicate]

EPPower [Ann] ExprPredicate

EPFuncAppl [Ann] Application

EPAppl [Ann] ExprPredicate ExprPredicate

EPDecor [Ann] ExprPredicate String

EPRename [Ann] ExprPredicate [RenamePair]

EPBindSel [Ann] ExprPredicate RefName

EPTupleSel [Ann] ExprPredicate String

EPTheta [Ann] ExprPredicate [String]

EPRef [Ann] RefName

EPGenRef [Ann] RefName [ExprPredicate]

EPEmptySet [Ann]

EPSet [Ann] [ExprPredicate]

EPSetComp [Ann] SchemaText ExprPredicate

EPCharSetComp [Ann] SchemaText

EPSchema [Ann] SchemaText

EPVarCons [Ann] NAME ExprPredicate

EPSchemaCons [Ann] ExprPredicate ExprPredicate

EPBindExt [Ann] [Declaration]

EPTupleExt [Ann] ExprPredicate [ExprPredicate]

EPCharMu [Ann] SchemaText

EPNum [Ann] String

EPRel [Ann] Relation

EPExprPred [Ann] ExprPredicate

EPTrue [Ann]

EPFalse [Ann]

data Expression

Expression

Constructors

NumExpr [Ann] String

ForallExpr [Ann] SchemaText Expression

ForallExprA [Ann] Expression Expression

ExistsExpr [Ann] SchemaText Expression

Exists1Expr [Ann] SchemaText Expression

Exists1ExprA [Ann] Expression Expression

LambdaExpr [Ann] SchemaText Expression

MuExpr [Ann] SchemaText Expression

MuExprA [Ann] Expression Expression

LocalDefExpr [Ann] [Declaration] Expression

EquivExpr [Ann] Expression Expression

ImplExpr [Ann] Expression Expression

AndExpr [Ann] Expression Expression

OrExpr [Ann] Expression Expression

NegExpr [Ann] Expression

CondExpr [Ann] Predicate Expression Expression

CompExpr [Ann] Expression Expression

PipeExpr [Ann] Expression Expression

HideExpr [Ann] Expression [DeclName]

HideExprA [Ann] Expression [NAME]

ProjExpr [Ann] Expression Expression

PreExpr [Ann] Expression

ProdExpr [Ann] [Expression]

PowerExpr [Ann] Expression

FuncApplExpr [Ann] Application

ApplExpr [Ann] Expression Expression

DecorExpr [Ann] Expression String

RenameExpr [Ann] Expression [RenamePair]

BindSelExpr [Ann] Expression RefName

BindSelExprA [Ann] Expression NAME

TupleSelExpr [Ann] Expression String

ThetaExpr [Ann] Expression [String]

RefExpr [Ann] RefName

RefExprA [Ann] NAME

GenRefExpr [Ann] RefName [Expression]

GenRefExprA [Ann] NAME [Expression]

EmptySetExpr [Ann]

SetExpr [Ann] [Expression]

SetCompExpr [Ann] SchemaText Expression

SetCompExprA [Ann] Expression Expression

CharSetCompExpr [Ann] SchemaText

SchemaExpr [Ann] SchemaText

VarConsExpr [Ann] NAME Expression

SchemaConsExpr [Ann] Expression Predicate

BindExtExpr [Ann] [Declaration]

BindExtExprA [Ann] [(NAME, Expression)]

TupleExtExpr [Ann] Expression [Expression]

CharMuExpr [Ann] SchemaText

NullExpr [Ann]

data DeclName

DeclName

Constructors

DName [Ann] String

OpName [Ann] OpName

data SchemaText

Schema text

Constructors

SchemaTextEmpty [Ann]
Empty Schema text

SchemaTextDecl [Ann] DeclPart

SchemaTextPred [Ann] Predicate

EPSchemaTextPred [Ann] ExprPredicate

SchemaText [Ann] DeclPart Predicate

EPSchemaText [Ann] DeclPart ExprPredicate

data DeclPart

DeclPart

Constructors

DeclPart [Ann] [Declaration]

data Declaration

Declaration

Constructors

Declaration [Ann] [DeclName] Expression

EPDeclaration [Ann] [DeclName] ExprPredicate

AbbrDecl [Ann] DeclName Expression

EPAbbrDecl [Ann] DeclName ExprPredicate

ExprDecl [Ann] Expression

EPExprDecl [Ann] ExprPredicate

Declaration1 [Ann] String Decl1

Declaration2 [Ann] OpName Decl1

data Decl1

Decl1

Predicate

Constructors

BasicDecl1 [Ann] [DeclName] Expression

EPBasicDecl1 [Ann] [DeclName] ExprPredicate

AbbrDecl1 [Ann] Expression

EPAbbrDecl1 [Ann] ExprPredicate

ExprDecl1 [Ann]

data Predicate

Constructors

AndPred [Ann] Predicate Predicate

ForallPred [Ann] SchemaText Predicate

ForallPredA [Ann] Expression Predicate

ExistsPred [Ann] SchemaText Predicate

Exists1Pred [Ann] SchemaText Predicate

Exists1PredA [Ann] Expression Predicate

EquivPred [Ann] Predicate Predicate

ImplPred [Ann] Predicate Predicate

OrPred [Ann] Predicate Predicate

NegPred [Ann] Predicate

RelPred [Ann] Relation

ExprPred [Ann] Expression

TruePred [Ann]

FalsePred [Ann]

MemPred [Ann] Expression Expression

data Formals

Constructors

Formals [Ann] [String]

data Freetype

Constructors

Freetype [Ann] NAME [Branch]

data Branch

Constructors

ConstantBranch [Ann] DeclName

ConstructorBranch [Ann] DeclName Expression

EPConstructorBranch [Ann] DeclName ExprPredicate

ConstantBranchA [Ann] NAME

ConstructorBranchA [Ann] NAME Expression

data OperatorTemplate

Constructors

RelationTemplate [Ann] Template

FunctionTemplate [Ann] CategoryTemplate

GenericTemplate [Ann] CategoryTemplate

data CategoryTemplate

Constructors

PrefixCatTemplate [Ann] PrefixTemplate

PostfixCatTemplate [Ann] PostfixTemplate

InfixCatTemplate [Ann] Prec Assoc InfixTemplate

NofixCatTemplate [Ann] NofixTemplate

data Prec

Constructors

Prec [Ann] String

data Assoc

Constructors

LeftAssoc [Ann]

RightAssoc [Ann]

data Template

Constructors

TPrefixTemplate [Ann] PrefixTemplate

TPostfixTemplate [Ann] PostfixTemplate

TInfixTemplate [Ann] InfixTemplate

TNofixTemplate [Ann] NofixTemplate

data PrefixTemplate

Constructors

PrefixTemplate [Ann] PrefixName

PowerPrefixTemplate [Ann]

data PostfixTemplate

Constructors

PostfixTemplate [Ann] PostfixName

data InfixTemplate

Constructors

InfixTemplate [Ann] InfixName

data NofixTemplate

Constructors

NofixTemplate [Ann] NofixName

data RefName

Constructors

RName [Ann] String

OpRName [Ann] OpName

data OpName

Constructors

PrefixOpName [Ann] PrefixName

PostfixOpName [Ann] PostfixName

InfixOpName [Ann] InfixName

NofixOpName [Ann] NofixName

data PrefixName

Constructors

PrePrefixName [Ann] String

PrePPrefixName [Ann] String

LPrefixName [Ann] String [ESSS] ERESRE

LPPrefixName [Ann] String [ESSS] EREPSREP

data PostfixName

Constructors

PostPostfixName [Ann] String

PostPPostfixName [Ann] String

ELPostfixName [Ann] String [ESSS] ERSR

ELPPostfixName [Ann] String [ESSS] ERPSRP

data InfixName

Constructors

InInfixName [Ann] String

InPInfixName [Ann] String

ELInfixName [Ann] String [ESSS] ERESRE

ELPInfixName [Ann] String [ESSS] EREPSREP

data NofixName

Constructors

LNofixName [Ann] String [ESSS] ERSR

LPNofixName [Ann] String [ESSS] ERPSRP

data ESSS

ES or SS operator name

For example, "function 1 leftassoc (_ el _ es1 _ es2 _ ere _)" contains

ES ann es1

ES ann es2

Constructors

ES [Ann] String

SS [Ann] String

data ESSS_N

ES or SS generic name

For example, "(X el Y es1 Z es2 U ere V)" contains

ES_N ann Y es1

ES_N ann Z es2

Constructors

ES_N [Ann] String String

SS_N [Ann] String String

data ESSS_E

ES or SS expression

For example, "(e1 el e2 es1 e3 es2 e4 ere e5)" contains

ES_E ann e2 es1

ES_E ann e3 es2

Constructors

ES_E [Ann] Expression String

EPES_E [Ann] ExprPredicate String

SS_E [Ann] [Expression] String

EPSS_E [Ann] [ExprPredicate] String

data ERESRE

Constructors

ERE [Ann] String

SRE [Ann] String

data ERESRE_N

Constructors

ERE_N [Ann] String String

SRE_N [Ann] String String

data ERESRE_E

Constructors

ERE_E [Ann] Expression String

EPERE_E [Ann] ExprPredicate String

SRE_E [Ann] [Expression] String

EPSRE_E [Ann] [ExprPredicate] String

data EREPSREP

Constructors

EREP [Ann] String

SREP [Ann] String

data EREPSREP_E

Constructors

EREP_E [Ann] Expression String

EPEREP_E [Ann] ExprPredicate String

SREP_E [Ann] [Expression] String

EPSREP_E [Ann] [ExprPredicate] String

data ERSR

Constructors

ER [Ann] String

SR [Ann] String

data ERSR_N

Constructors

ER_N [Ann] String String

SR_N [Ann] String String

data ERSR_E

Constructors

ER_E [Ann] Expression String

EPER_E [Ann] ExprPredicate String

SR_E [Ann] [Expression] String

EPSR_E [Ann] [ExprPredicate] String

data ERPSRP

Constructors

ERP [Ann] String

SRP [Ann] String

data ERPSRP_E

Constructors

ERP_E [Ann] Expression String

EPERP_E [Ann] ExprPredicate String

SRP_E [Ann] [Expression] String

EPSRP_E [Ann] [ExprPredicate] String

data GenName

Constructors

PrefixGenName [Ann] PrefixGenName

PostfixGenName [Ann] PostfixGenName

InfixGenName [Ann] InfixGenName

NofixGenName [Ann] NofixGenName

data PrefixGenName

Constructors

PrePrefixGenName [Ann] String String

LPrefixGenName [Ann] String [ESSS_N] ERESRE_N String

data PostfixGenName

Constructors

PostPostfixGenName [Ann] String String

ELPostfixGenName [Ann] String String [ESSS_N] ERSR_N

data InfixGenName

Constructors

InInfixGenName [Ann] String String String

ELInfixGenName [Ann] String String [ESSS_N] ERESRE_N String

data NofixGenName

Constructors

LNofixGenName [Ann] String [ESSS_N] ERSR_N

data Relation

Constructors

PrefixRel [Ann] PrefixRel

PostfixRel [Ann] PostfixRel

InfixRel [Ann] InfixRel

NofixRel [Ann] NofixRel

data PrefixRel

Constructors

PrePPrefixRel [Ann] String Expression

EPPrePPrefixRel [Ann] String ExprPredicate

LPPrefixRel [Ann] String [ESSS_E] EREPSREP_E Expression

EPLPPrefixRel [Ann] String [ESSS_E] EREPSREP_E ExprPredicate

data PostfixRel

Constructors

PostPPostfixRel [Ann] Expression String

EPPostPPostfixRel [Ann] ExprPredicate String

ELPPostfixRel [Ann] Expression String [ESSS_E] ERPSRP_E

EPELPPostfixRel [Ann] ExprPredicate String [ESSS_E] ERPSRP_E

data InfixRel

Constructors

InInfixRel [Ann] Expression [IP_E]

EPInInfixRel [Ann] ExprPredicate [IP_E]

ELPInfixRel [Ann] Expression String [ESSS_E] EREPSREP_E Expression

EPELPInfixRel [Ann] ExprPredicate String [ESSS_E] EREPSREP_E ExprPredicate

data NofixRel

Constructors

LPNofixRel [Ann] String [ESSS_E] ERPSRP_E

data IP_E

Constructors

IP_E [Ann] String Expression

EPIP_E [Ann] String ExprPredicate

data Application

Constructors

PrefixApp [Ann] PrefixApp

PostfixApp [Ann] PostfixApp

InfixApp [Ann] InfixApp

NofixApp [Ann] NofixApp

data PrefixApp

Constructors

PrePrefixApp [Ann] String Expression

EPPrePrefixApp [Ann] String ExprPredicate

LPrefixApp [Ann] String [ESSS_E] ERESRE_E Expression

EPLPrefixApp [Ann] String [ESSS_E] ERESRE_E ExprPredicate

data PostfixApp

Constructors

PostPostfixApp [Ann] Expression String

EPPostPostfixApp [Ann] ExprPredicate String

ELPostfixApp [Ann] Expression String [ESSS_E] ERSR_E

EPELPostfixApp [Ann] ExprPredicate String [ESSS_E] ERSR_E

data InfixApp

Constructors

InInfixApp [Ann] Expression String Expression

EPInInfixApp [Ann] ExprPredicate String ExprPredicate

ELInfixApp [Ann] Expression String [ESSS_E] ERESRE_E Expression

EPELInfixApp [Ann] ExprPredicate String [ESSS_E] ERESRE_E ExprPredicate

data NofixApp

Constructors

LNofixApp [Ann] String [ESSS_E] ERSR_E

EPLNofixApp [Ann] String [ESSS_E] ERSR_E

data ChannelDecl

Constructors

ChannelDecl [Ann] SimpleCDecl

data SimpleCDecl

Constructors

SyncSimpleCDecl [Ann] [String]

TypedSimpleCDecl [Ann] [String] Expression

GenericTypedSimpleCDecl [Ann] Formals [String] Expression

SimpleCDecl2 [Ann] [String] SimpleCDecl2

data SimpleCDecl2

Constructors

SyncSimpleCDecl2 [Ann]

TypedSimpleCDecl2 [Ann] Expression

data ChannelFromDecl

Constructors

ChannelFromDecl [Ann] FromCDecl

data FromCDecl

Constructors

RefFromCDecl [Ann] Expression

RefDecorFromCDecl [Ann] Expression

data ChannelSetDecl

Constructors

ChannelSetDecl [Ann] NAME CSExp

data CSExp

Constructors

CSExpEmpty [Ann]

CSExpExt [Ann] [Expression]

CSExpRef [Ann] Expression

CSExpUnion [Ann] CSExp CSExp

CSExpInter [Ann] CSExp CSExp

CSExpDiff [Ann] CSExp CSExp

data ProcDef

Constructors

ParamProc [Ann] ProcDef

IndexedProc [Ann] ProcDef

getPrec :: CategoryTemplate -> String

getAssoc :: CategoryTemplate -> String

getOpNameFromCatTemplate :: CategoryTemplate -> String

getOpNameFromTemplate :: Template -> String

getOpNameFromPrefixTemplate :: PrefixTemplate -> String

getOpNameFromPostfixTemplate :: PostfixTemplate -> String

getOpNameFromInfixTemplate :: InfixTemplate -> String

getOpNameFromNofixTemplate :: NofixTemplate -> String