Language.ISOZ.Common.Ann

Description

Annotation type and constructors (see clause 10 of ISO Z Standard)

Documentation

data Posn

position (offset, line, column)

Constructors

Pn !Int !Int !Int

data Ann

Annotation type.

Constructors

PosAnn Posn
position annotation

SectTypeEnvAnn SectTypeEnv
section type environment annotation

SigAnn Sig
signature annotation

TypeAnn Type
type annotation

TypeVarAnn [String]
type variable annotation (a list of temporarily allocated type variables) | an annotation to store the operator table for all operators defined in templates, usually it only occurs in the annotation list of Specification

OpMapAnn (Map String String)

SchNameMapAnn (Map String Bool)
a map of defined schemas

ExprInParAnn
indicate this expression is the outermost expression within a parenthesis (such as (a+b))

ParserParDepthAnn Int
the parenthesis depth of a ProdExpr, for example, in ((A x B) x C): the depth of the first x is equal to that of the second + 1

isSameTypeAnn :: Ann -> Ann -> Bool

Check if two annotations have the same type or not.

addAnnToAnnListHelper :: Ann -> [Ann] -> [Ann]

Add an Ann to a list of Ann. If there has been a same type of annotation, just override it.

Assumption: no duplicate type of Ann in the list.

extractLineCol :: Posn -> (Int, Int)

Extract line number and column number from position information

getPosnFromListAnns :: [Ann] -> Posn

Get Posn from a list of annotations

opGlueSTROKE :: [Char]

Operator glue stroke ⋈

genTypeSTROKE :: [Char]

Generic type stroke ♠

givenTypeSTROKE :: [Char]

Given type stroke ♡

getAndRemoveTypeVarAnn :: [Ann] -> [Ann] -> ([String], [Ann])

Remove an Ann from a list

the first parameter: the original list of Ann

the second parameter: the visited list of Ann

getOpMapAnnFromListAnns :: [Ann] -> Map String String

Get operator map annotation from a list of annotations

removeOpMapFromListAnns :: [Ann] -> [Ann]

Remove a SchNameMapAnn annotation in a list of annotations

getSchNameMapAnnFromListAnns :: [Ann] -> Map String Bool

Get schema name map annotation from a list of annotations

removeSchNameMapFromListAnns :: [Ann] -> [Ann]

Remove a SchNameMapAnn annotation in a list of annotations

isExprInParFromListAnns :: [Ann] -> Bool

Check if there is an ExprInPar annotation in a list of annotations

removeExprInParFromListAnns :: [Ann] -> [Ann]

Remove an ExprInPar annotation in a list of annotations

getParserParDepthFromListAnns :: [Ann] -> (Bool, Int)

Get ParserParDepthAnn from a list of annotations

getParserParDepthFromListAnnsN :: [Ann] -> Int

Get the depth of ParserParDepthAnn from a list of annotations

removeParserParDepthFromListAnns :: [Ann] -> [Ann]

Remove ParserParDepthAnn from a list of annotations