Language.ISOZ.Common.Error

Description

Error messages of the parser.

Documentation

data ErrorType

All error categories

Constructors

ErrSecSolve ErrSecSolve
Errors during section solving

ErrTex2Uni ErrTex2Uni
Errors during translation of Latex to Unicode

ErrUniLexer ErrUniLexer
Errors duing lexing of unicode

ErrParser ErrParser
Errors duing parsing

ErrTypeCheck ErrTypeCheck
Errors duing type checking

data ErrSecSolve

Errors during sections solving of Latex specifications

Constructors

ESecLexer String
Lexing error of sections environment

ESecNoFile String
file doesn't exist

ESecNoEnv String
Env variable doesn't exist

data ErrTex2Uni

Errors during translation of Latex specifications to Unicode

Constructors

ET2ULexer String
Lexing error of sections environment

ET2UCmdNotDef String
a cmd is not defined

ET2UCmdEmpty String

ET2UCmdRedef String

ET2UUnbalStruct String

ET2UResSymb String

ET2UNoEndEnv String

ET2UInternal String

data ErrUniLexer

Constructors

EUniLexer String

data ErrParser

Constructors

EParseSyntaxError String
general syntax error

EParseNotASchema String
not a schema expression error

EParseEPNotExpr String
ExprPredicate cannot be expression

EParseEPNotPred String
ExprPredicate cannot be predicate

EParseOpNotDef String
Operator not defined

EParseOpInvalidPrec String
Invalid operator precedence

EParseDiffAssocUsed String
Different associativities used for operators with the same precedence

data ErrTypeCheck

Constructors

ETypeCDupSect String
duplicate/conflict section names

ETypeCUndefSect String
undefined parent or section

ETypeCDupName String
duplicate/conflict name definition

ETypeCGenericUninst String
not all generic variables are instantiated

ETypeCNotSchType String
not all generic variables are instantiated

ETypeCTypeMismatch String

ETypeCTypeMismatch1 String
expect: xxx Actual: yyy

ETypeCTypeMismatch2 String

ETypeCUndefName String
undefined name

ETypeCMismatchNoOfPara String

ETypeCNotGenericType String
not a generic type error | ETypeCNotSetExpr Expression -- ^ not a set expression

ETypeCNotAllSameType String [Expression]
not all expressions in the list have the same type

ETypeCIdxOutOfRange String

ETypeCNameNotInSig String

ETypeCSigIncomp String
incompatible signature

ETypeCNameExistInSig String
the new name exists in current signature

ETypeCTypeVarInCarrier String
type variable occurs in carrier

err2Msg :: ErrorType -> String