Language.ISOZ.Common.Type

Description

annotated syntax (see clause 10.2).

Documentation

data SectTypeEnv

section type environment

Constructors

SectTypeEnv (Map String (String, Type))

type SectEnv = Map String SectTypeEnv

section environment (map from section names to SectTypeEnv)

data Type

Type (see clause 10)

Constructors

Type Type2
Type2

GenType [String] Type2
Generic Type2 ([X] GIVEN)

GenType2 [String] Type2 Type2
Generic Type2 ([X] GIVEN, GIVEN)

data Type2

Type2 (see clause 10.2)

Constructors

GIVEN String
given type

GENTYPE String
generic parameter type

PowerType Type2
powerset type

ProdType [Type2]
Cartesian product type

SchType Sig
schema type

TypeVar String
type variable | NumType String

type Sig

Arguments

= Map String Type
a map from names to their types deriving (Eq, Ord, Show, Read)

signature

getSectTypeEnvMap :: SectTypeEnv -> Map String (String, Type)

get section type environment's map

sectTypeEnvToTypeEnv :: SectTypeEnv -> [(String, Type)]

convert a SectTypeEnv to a list of pair from a name to its type

extractTypeEnvFromSectTypeEnv :: (String, (String, Type)) -> (String, Type)

extract a type environment from section type environment

typeToType2 :: Type -> Type2

Get Type2 of a Type