Language.ISOZ.Common.TypeCMonad

Description

A state monad to maintain information for parsing and typechecking.

Documentation

data TypeState

the state enclosed

Constructors

TypeState

Fields

env :: Map String Type

type environment, from NAME to a Type

sectEnv :: Map String SectTypeEnv

section environment, from section name to SectTypeEnv

curSectName :: String

current section name

typeVarList :: Map String Type2

a list of pair from a type variable to a Type2. Initially, its type is equal to (TypeVar a). Therefore, if a type variable (α) has its type equal to (TypeVar α). It means it is not instantiated

opNameFuncGen :: Map String Bool

true for function

newtype TypeC a

Constructors

TypeC

Fields

unTypeC :: TypeState -> Either String (TypeState, a)

unTypeC :: TypeC a -> TypeState -> Either String (TypeState, a). A function to unwrap (state, result) from a TypeC. unTypeC (TypeC a) oldState . Return (Left str) or Right (NewState, result)

runTypeC :: TypeC a -> Either String a

getEnvTypeC :: TypeC (Map String Type)

get Env from State

setEnvTypeC :: Map String Type -> TypeC ()

set Env to State

getSectEnvTypeC :: TypeC (Map String SectTypeEnv)

get SectEnv from State

setSectEnvTypeC :: Map String SectTypeEnv -> TypeC ()

set SectEnv to State

getCurSectName :: TypeC String

get curSectName from State

setCurSectName :: String -> TypeC ()

set curSectName to State

getOpNameFuncGen :: TypeC (Map String Bool)

get opNameFuncGen from State

setOpNameFuncGen :: Map String Bool -> TypeC ()

set opNameFuncGen from State

typeError :: String -> TypeC a

runDo :: TypeC String

getTypeVarList :: TypeC (Map String Type2)

get typeVarList from State

setTypeVarList :: Map String Type2 -> TypeC ()

set typeVarList to State

clearTypeVarList :: TypeC ()

clear all type variables

allocAddANewTypeVar :: TypeC String

allocAddANewTypeVar will add a variable to the map - If there are (n-1) variables in the typeVarList, then the new type variable name - will be ("αn", TypeVar "αn") - and return "αn"

getTypeVarType2 :: String -> TypeC (Bool, Type2)

check if a type variable is in the typeVarList or not - return - (True, t) - in the typeVarList and return its bound type - (False, TypeVar v) - not in the typeVarList

typeVarsInAType2 :: Type2 -> [String]

get all type variables which occur in a Type2

isATypeVarInstantiated :: String -> Map String Type2 -> Bool

check if a type variable is instantiated or not

isTypeVarsInstantiated :: [String] -> Map String Type2 -> Bool

check if a list of type variables is instantiated or not

isAllGenInsted :: TypeC Bool

check if all generic types are instantiated

instATypeVar :: String -> Type2 -> TypeC ()

instantiate a type variable (v) to a type (t) at the same time, bind all type variables whose type2 is linked to this type variable, to this type2

updateAllType2BoundThisVar :: String -> Type2 -> TypeC ()

Update all entries in typeVarList that are bound to the type variable v by its type t

replaceAllTypeVarsInAType2 :: Type2 -> TypeC (Either String Type2)

Replace all type variables in a Type2 by its instantiated types