Language.ISOZ.Lexer.ISOZLexer

Description

Lexis (clause 6 and clause 7).

Documentation

quickIndex :: Array Int (AlexAcc (Any *)) -> Int -> AlexAcc (Any *)

data AlexReturn a

Constructors

AlexEOF

AlexError !AlexInput

AlexSkip !AlexInput !Int

AlexToken !AlexInput !Int a

alexScan :: AlexInput -> Int -> AlexReturn (AlexInput -> Int -> Alex Token)

alexScanUser :: Any * -> AlexInput -> Int -> AlexReturn (AlexInput -> Int -> Alex Token)

alex_scan_tkn :: Any * -> AlexInput -> Int# -> AlexInput -> Int# -> AlexLastAcc -> (AlexLastAcc, AlexInput)

data AlexLastAcc

Constructors

AlexNone

AlexLastAcc !Int !AlexInput !Int

AlexLastSkip !AlexInput !Int

data AlexAcc user

Constructors

AlexAccNone

AlexAcc Int

AlexAccSkip

AlexAccPred Int (AlexAccPred user) (AlexAcc user)

AlexAccSkipPred (AlexAccPred user) (AlexAcc user)

type AlexAccPred user = user -> AlexInput -> Int -> AlexInput -> Bool

alexAndPred :: (t -> t1 -> t2 -> t3 -> Bool) -> (t -> t1 -> t2 -> t3 -> Bool) -> t -> t1 -> t2 -> t3 -> Bool

alexPrevCharIs :: Char -> t -> AlexInput -> t1 -> t2 -> Bool

alexPrevCharMatches :: (Char -> t3) -> t -> AlexInput -> t1 -> t2 -> t3

alexPrevCharIsOneOf :: Array Char e -> t -> AlexInput -> t1 -> t2 -> e

alexRightContext :: Int -> AlexAccPred (Any *)

data Tok a

token types

Constructors

T_NIL
Nil

T_LPARENT
(

T_RPARENT
)

T_LBRACKET
[

T_LBRACKET2
[ before a list of declaration, such as "[x : T, y: Ty"

T_LBRACKET3
[ before a list of rename pair, such as "[xx1, yy1"

T_RBRACKET
]

T_RBRACKET2
] before a list of declaration, such as "]x : T, y: Ty"

T_LBRACE
'{'

T_LBRACE2
'{' before a list of declaration, such as "{x : T, y: Ty"

T_RBRACE
'}'

T_LBLOT
⟨|

T_RBLOT
|⟩

T_LDATA
⟨⟨

T_RDATA
⟩⟩

T_LCHANSET

T_RCHANSET

T_LCIRCINDEX

T_RCIRCINDEX

T_LCIRCGUARD

T_RCIRCGUARD

T_LCIRCRENAME

T_RCIRCRENAME

T_LPAR

T_RPAR

T_LINTER

T_RINTER

T_NUMERAL

T_STROKE
Strokes: ’′’| ’!’| ’?’ | _{n}

T_ZED

T_ZED_SEC
specific for ZED before "sections"

T_AX

T_AX2

T_SCH

T_CONJ

T_GENAX

T_GENSCH

T_GENCONJ

T_END

T_NL

T_NL2

T_WORD

T_DECORWORD

T_KEYWORD !a !Int

T_KEYWORD2 !a !Int
add T_START_SCHEMATEXT before it operator and user defined operator

T_OPERATOR

T_PREP

T_PRE

T_POSTP

T_POST

T_IP

T_I

T_LP1
LP for Prefix

T_L1
L for Prefix

T_LP4
LP for Nofix

T_L4
L for Nofix

T_ELP2
ELP for Postfix

T_EL2
EL for Postfix

T_ELP3
ELP for Infix

T_EL3
EL for Infix

T_ERP2
ERP for Postfix

T_ER2
ER for Postfix

T_ERP4
ERP for Nofix

T_ER4
ER for Nofix

T_SRP2
SRP for Postfix

T_SR2
SR for Postfix

T_SRP4
SRP for Nofix

T_SR4
SR for Nofix

T_EREP1
EREP for Prefix

T_ERE1
ERE for Prefix

T_EREP3
EREP for Infix

T_ERE3
ERE for Infix

T_SREP1
SREP for Prefix

T_SRE1
SRE for Prefix

T_SREP3
SREP for Infix

T_SRE3
SRE for Infix

T_ES

T_SS

T_NAME

T_NAME2

T_START_SCHEMATEXT
special for placemark NLCHAR (7.5) newlines

data Token

Token

Constructors

PT AlexPosn (Tok String) (Maybe String)

Err AlexPosn

EOF

tokenPos :: [Token] -> String

Token to line number

tokenPosn :: Token -> AlexPosn

Token to Posn

tokenLineCol :: Token -> (Int, Int)

posLineCol :: AlexPosn -> (Int, Int)

get a pair of line number and column number

mkPosToken :: Token -> ((Int, Int), String)

prToken :: Token -> String

get a Token's string

getListOfOpTok :: [Tok String]

get a list of operator tokens

mkL :: Tok String -> AlexInput -> Int -> Alex Token

make a Token. If it is an operator token, add it to nameTokenMap and then we can use it determine if a Token is an operator or not

isSubscript :: String -> Int

check if a string is a subscribe "_", unicode "↘ ↖"

subIndex

Arguments

:: String
string to be processed

-> Int
index of char in processing

-> Int
index of SE

-> Int
return the index of SE

determine if a string is a subscript (such as "abc_1", "a_{12}") or not. If it is, return the index of SE

If index = 0, it means it is not a subscript. For example,

1. "a_1" => 1

2. "_1" => 0

3. "a_12" => 0

mkKeyword :: AlexInput -> Int -> Alex Token

if it is not a keyword and a T_NAME instead, check nameTokenMap map if this name has been linked to a Token or not. If so, just use this linked Token for this NAME. Otherwise, it is a T_NAME token

mkKeyword2 :: AlexInput -> Int -> Alex Token

if it is not a keyword and a T_NAME instead, check nameTokenMap map if this name has been linked to a Token or not. If so, just use this linked Token for this NAME. Otherwise, it is a T_NAME token

enterOpRelationMode :: AlexInput -> Int -> Alex Token

enter relation mode

enterOpFunctionMode :: AlexInput -> Int -> Alex Token

enter function mode

enterOpGenericMode :: AlexInput -> Int -> Alex Token

enter generic mode

mkOpKeyword :: (Tok String, Tok String) -> AlexInput -> Int -> Alex Token

make operator keyword according to current op mode (relationfunctiongeneric)

enterPrefixName2 :: AlexInput -> Int -> Alex Token

just before "_)"

enterPrefixName :: AlexInput -> Int -> Alex Token

(_ ES | ,, SS)* (_, ERE | ,, SRE) _)

enterPostfixName2 :: AlexInput -> Int -> Alex Token

just before ")"

enterPostfixName :: AlexInput -> Int -> Alex Token

EL , {_ , ES| ,, SS } , (_, ER| ,, SR )

enterInfixName2 :: AlexInput -> Int -> Alex Token

just before "_)"

enterInfixName :: AlexInput -> Int -> Alex Token

EL , {_ , ES| ,, SS } , (_, ERE| ,, SRE ) _

enterNofixName2 :: AlexInput -> Int -> Alex Token

just before ")"

enterNofixName :: AlexInput -> Int -> Alex Token

L , {_ , ES| ,, SS } , (_, ER| ,, SR )

data BTree a where

Keyword search binary tree

Constructors

N :: Ord a => BTree a

B :: Ord a => a -> Tok a -> BTree a -> BTree a -> BTree a

reservedDecorwords :: Tok String -> String -> Tok String

search reserved keywords to check if it is a reserved word or not

resWords :: BTree String

return a BTree for reserved words

unescapeInitTail :: String -> String

buildBST :: (Ord a, Eq a) => [(a, Int)] -> BTree a

build the balanced binary search tree, and input list should be sorted in advance, such as buildBST [("a", 1), ("b", 2), ...]

middle :: [a] -> (a, Int)

a helper function that gives the element in the middle and the index of it

appendNumbering :: [String] -> Int -> [(String, Int)]

append number to each String in the list to get a pair of (String, Int)

For example, 'appendNumbering ["a", "b"] 1' will get [("a",1),("b",2)]

alexEOF :: Alex Token

data OpMode

user defined operator template mode (relation, function and generic template) or not

Constructors

OpRelation

OpFunction

OpGeneric

OpOther

data AlexUserState

User State

Constructors

AlexUserState

Fields

lexerCommentDepth :: Int

lexerZedDepth :: Int

ZED depths

opTemplate :: OpMode

operator template mode | last two tokens (last-1, last), for user defined operators

lastToken :: (Tok String, Tok String)

nameTokenMap :: Map String (Tok String)

a (key, value) hash map from a T_NAME to Tok. It is used to store the visited name and its Tok. So in the later stage, if a same name is encountered, we use this Tok directly. This is expecially for operator template

schemaExprMap :: Map String Bool

a (schemaName, bool) hash map from a name to a boolean to indicate if it is schema or not

opTable :: Map String String

An operator table to store all operators defined in templates. A pair from the name of operator to its related infomation including (relation|function|generic : prec : assoc).

For example: - ("+", "function:30:leftassoc") - ("≠", "relation") - ("ℙ", "generic::")

parserParDepth :: Int

Depth of expressions in LPARENT and RPARENT, and used in parsing stage to cope with Catersian product like (X x Y x Z) ==== should be ====> ProdExpr [X, Y, Z] where the depthes of both x are equal ((X x Y) x Z) ==== should be ====> ProdExpr [(ProdExpr [X, Y]), Z] where the depth of the first x is equal to that of the second + 1 (X x Y op U x Z) ==== should be ====> ProdExpr [X, (ApplExpr InfixApp Y op U), Z] where the depth of the first x is equal to that of the second and op binds more tightly than x

alexInitUserState :: AlexUserState

initialisation of user states

Notes: the Catersian product has the precedence 8 according to p42 of ISO Standard Z

getLexerZedDepth :: Alex Int

get ZedDepth

setLexerZedDepth :: Int -> Alex ()

set ZedDepth

enterNewZed :: Tok String -> AlexInput -> Int -> Alex Token

enter a new ZED environment (ZedDepth = 1)

embedZed :: Tok String -> AlexInput -> Int -> Alex Token

enter an embedded ZED environment (ZedDepth = ZedDepth + 1)

unembedZed :: Tok String -> AlexInput -> Int -> Alex Token

exit an embedded ZED environment (ZedDepth = ZedDepth - 1)

getLexerCommentDepth :: Alex Int

setLexerCommentDepth :: Int -> Alex ()

enterNewComment :: Any * -> Any * -> Alex Token

embedComment :: Any * -> Any * -> Alex Token

unembedComment :: Any * -> Any * -> Alex Token

state_initial :: Int

getOpMode :: Alex OpMode

get operator mode

setOpMode :: OpMode -> Alex ()

set operator mode

getLast2Token :: Alex (Tok String, Tok String)

get last two Tokens

getLastToken :: Alex (Tok String)

get last Token

getLast1Token :: Alex (Tok String)

get the second last Token

setLast2Token :: (Tok String, Tok String) -> Alex ()

set last two tokens

setLastToken :: Token -> Alex ()

set last token

getNameTokenMap :: Alex (Map String (Tok String))

get nameTokenMap

setNameTokenMap :: Map String (Tok String) -> Alex ()

set nameTokenMap

getSchemaExprMap :: Alex (Map String Bool)

get schemaExprMap

setSchemaExprMap :: Map String Bool -> Alex ()

set schemaExprMap

getOpTable :: Alex (Map String String)

get OpTable

setOpTable :: Map String String -> Alex ()

set OpTable

getParserParDepth :: Alex Int

get parserParDepth

setParserParDepth :: Int -> Alex ()

set parserParDepth

showPosn :: AlexPosn -> String

lexerError :: String -> Alex a

errMsg :: String -> AlexInput -> Int -> Alex a

isNL :: Token -> Bool

isHardNL :: Token -> Token -> Bool

check if it is a soft NL or hard NL.

a NLCHAR is soft only if it is checked to be soft from both the token before and the token after

scanner :: String -> Either String [Token]

body :: Int

optemp :: Int

optempassoc :: Int

optempprec :: Int

relationIn :: Int

relationNo :: Int

relationPost :: Int

relationPre :: Int

alex_action_2 :: AlexAction Token

alex_action_3 :: t -> t1 -> Alex a

alex_action_4 :: AlexAction Token

alex_action_5 :: AlexAction Token

alex_action_6 :: AlexAction Token

alex_action_7 :: AlexAction Token

alex_action_8 :: AlexAction Token

alex_action_9 :: AlexAction Token

alex_action_10 :: AlexAction Token

alex_action_11 :: AlexAction Token

alex_action_12 :: t -> t1 -> Alex a

alex_action_15 :: AlexInput -> Int -> Alex Token

alex_action_16 :: AlexInput -> Int -> Alex Token

alex_action_17 :: AlexInput -> Int -> Alex Token

alex_action_18 :: AlexInput -> Int -> Alex Token

alex_action_19 :: AlexInput -> Int -> Alex Token

alex_action_20 :: AlexInput -> Int -> Alex Token

alex_action_21 :: AlexInput -> Int -> Alex Token

alex_action_22 :: AlexInput -> Int -> Alex Token

alex_action_23 :: AlexInput -> Int -> Alex Token

alex_action_24 :: AlexInput -> Int -> Alex Token

alex_action_25 :: AlexInput -> Int -> Alex Token

alex_action_26 :: AlexInput -> Int -> Alex Token

alex_action_27 :: AlexInput -> Int -> Alex Token

alex_action_28 :: AlexInput -> Int -> Alex Token

alex_action_29 :: AlexInput -> Int -> Alex Token

alex_action_30 :: AlexInput -> Int -> Alex Token

alex_action_31 :: AlexInput -> Int -> Alex Token

alex_action_32 :: AlexInput -> Int -> Alex Token

alex_action_33 :: AlexInput -> Int -> Alex Token

alex_action_34 :: AlexInput -> Int -> Alex Token

alex_action_35 :: AlexInput -> Int -> Alex Token

alex_action_36 :: AlexInput -> Int -> Alex Token

alex_action_37 :: AlexInput -> Int -> Alex Token

alex_action_38 :: AlexInput -> Int -> Alex Token

alex_action_39 :: AlexInput -> Int -> Alex Token

alex_action_40 :: AlexInput -> Int -> Alex Token

alex_action_41 :: AlexInput -> Int -> Alex Token

alex_action_42 :: AlexInput -> Int -> Alex Token

alex_action_43 :: AlexInput -> Int -> Alex Token

alex_action_44 :: AlexInput -> Int -> Alex Token

alex_action_45 :: AlexInput -> Int -> Alex Token

alex_action_46 :: AlexInput -> Int -> Alex Token

alex_action_47 :: AlexInput -> Int -> Alex Token

alex_action_48 :: AlexInput -> Int -> Alex Token

alex_action_49 :: AlexInput -> Int -> Alex Token

alex_action_50 :: AlexInput -> Int -> Alex Token

alex_action_51 :: AlexInput -> Int -> Alex Token

alex_action_52 :: AlexAction Token

alex_action_53 :: AlexAction Token

alex_action_54 :: AlexAction Token

alex_action_55 :: AlexAction Token

alex_action_56 :: AlexAction Token

alex_action_57 :: AlexInput -> Int -> Alex Token

alex_action_58 :: AlexInput -> Int -> Alex Token

alex_action_59 :: AlexAction Token

alex_action_60 :: AlexAction Token

alex_action_61 :: AlexAction Token

alex_action_62 :: AlexAction Token

alex_action_63 :: AlexAction Token

alex_action_64 :: AlexAction Token

alex_action_65 :: AlexInput -> Int -> Alex a

alex_action_66 :: AlexInput -> Int -> Alex Token

alex_action_67 :: AlexInput -> Int -> Alex Token

alex_action_68 :: AlexInput -> Int -> Alex Token

alex_action_69 :: AlexInput -> Int -> Alex Token

alex_action_70 :: AlexAction Token

alex_action_71 :: AlexInput -> Int -> Alex Token

alex_action_72 :: AlexInput -> Int -> Alex Token

alex_action_73 :: AlexInput -> Int -> Alex Token

alex_action_74 :: AlexInput -> Int -> Alex Token

alex_action_75 :: AlexAction Token

alex_action_76 :: AlexInput -> Int -> Alex Token

alex_action_77 :: AlexInput -> Int -> Alex Token

alex_action_78 :: AlexInput -> Int -> Alex Token

alex_action_79 :: AlexInput -> Int -> Alex Token

alex_action_80 :: AlexAction Token

alex_action_81 :: AlexInput -> Int -> Alex Token

alex_action_82 :: AlexInput -> Int -> Alex Token

alex_action_83 :: AlexInput -> Int -> Alex Token

alex_action_84 :: AlexInput -> Int -> Alex Token

alex_action_85 :: AlexAction Token

data AlexAddr

Constructors

AlexA# Addr#

alexIndexInt16OffAddr :: AlexAddr -> Int# -> Int#

alexIndexInt32OffAddr :: AlexAddr -> Int# -> Int#

utf8Encode :: Char -> [Word8]

Encode a Haskell String to a list of Word8 values, in UTF8 format.

type Byte = Word8

type AlexInput = (AlexPosn, Char, [Byte], String)

ignorePendingBytes :: AlexInput -> AlexInput

alexInputPrevChar :: AlexInput -> Char

alexGetByte :: AlexInput -> Maybe (Byte, AlexInput)

data AlexPosn

Constructors

AlexPn !Int !Int !Int

alexStartPos :: AlexPosn

alexMove :: AlexPosn -> Char -> AlexPosn

data AlexState

Constructors

AlexState

Fields

alex_pos :: !AlexPosn

alex_inp :: String

alex_chr :: !Char

alex_bytes :: [Byte]

alex_scd :: !Int

alex_ust :: AlexUserState

runAlex :: String -> Alex a -> Either String a

newtype Alex a

Constructors

Alex

Fields

unAlex :: AlexState -> Either String (AlexState, a)

alexGetInput :: Alex AlexInput

alexSetInput :: AlexInput -> Alex ()

alexError :: String -> Alex a

alexGetStartCode :: Alex Int

alexSetStartCode :: Int -> Alex ()

alexGetUserState :: Alex AlexUserState

alexSetUserState :: AlexUserState -> Alex ()

alexMonadScan :: Alex Token

type AlexAction result = AlexInput -> Int -> Alex result

skip :: Any * -> Any * -> Alex Token

begin :: Int -> Any * -> Any * -> Alex Token

andBegin :: AlexAction result -> Int -> AlexAction result

token :: (AlexInput -> Int -> token) -> AlexAction token

alex_tab_size :: Int

alex_base :: AlexAddr

alex_table :: AlexAddr

alex_check :: AlexAddr

alex_deflt :: AlexAddr

alex_accept :: Array Int (AlexAcc (Any *))

alex_actions :: Array Int (AlexInput -> Int -> Alex Token)