Language.ISOZ.Parser.Mktuple

Description

Syntactical transformation to cope with "Clause 9: characteristic tuple" in ISO Z Standard.

Documentation

chartuple :: SchemaText -> Expression

Converts a schema text to an expression as stated in Clause 9.2

chartuple t = mktuple (charac t)

getListStrokes :: Expression -> [String]

Get a list of strokes from a DecorExpr. For example, s'' is parsed as (DecorExpr (DecorExpr (RefExpr s) ') ')

getListStrokes (DecorExpr (DecorExpr (RefExpr s) ') ') == ["'", "'"]

getExpr :: Expression -> Expression

Get the expression from a DecorExpr. For example, s'' is parsed as (DecorExpr (DecorExpr (RefExpr s) ') ')

getExpr (DecorExpr (DecorExpr (RefExpr s) ') ') ==  (DecorExpr (RefExpr s) ')

mktuple :: [Expression] -> Expression

Turn sequences of expressions enclosed between metalanguage brackets and, such as e2, e3 into an expression

Definition:

mktuple e = e

mktuple e2, e3 = (e1, e2, e3)

mktuple [e]              == e
mktuple [e1, e2, e3]     == (e1, e2, e3)

class Mktuple a where

A class to provide charac for schema texts, declarations, and declaration names

Specification:

charac (d1; ...; dn | p) = charac (d1; ...; dn)

charac (d1 ; ...; dn ) = charac d1 ^ ... ^ charac dn where n >= 1 (concatenation)

charac () = ⟨⟨| |⟩⟩

charac (i1, ..., in : e) = ⟨i1 , ..., in ⟩

charac (i == e) = ⟨i⟩

charac (e ∗) = ⟨θe∗⟩

Minimal complete definition

charac

Methods

charac :: a -> [Expression]

characList :: [a] -> [Expression]