Language.ISOZ.Parser.SynTransformExpr

Description

Syntactical transformation to cope with operator precedences and associativities for expressions.

It provides

synTraverseExprTree : traverse the syntax tree, particularly expressions

Documentation

synTraverseExprTree :: SynTraverseExpr a => a -> Either String a

Traverse the syntax tree.

class SynTraverseExpr a where

Minimal complete definition

traverseExpr

Methods

traverseExpr :: a -> OpTableType (Either String a)

traverseExprList :: [a] -> OpTableType (Either String [a])

compareOpPrecAssoc :: String -> String -> OpTableType (Either String Int)

Compare two infix operators' precedences and associativity

if prec(op1) > prec(op2), return 1 (op1 binds tighter than op2)

if prec(op1) < prec(op2), return 0 (op2 binds tighter than op1)

if prec(op1) = prec(op2), and

if assoc1 == assoc2, and

leftassoc return 1 (op1 binds tighter than op2)

rightassoc return 0 (op2 binds tighter than op1)

else return -1 (error ("different associativities shouldn't be used in the same precedence!"))

getPrecAndAssoc :: String -> OpTableType (Either String (Int, String))

Return (precedence, associativity) of an operator

getPrecAndAssoc "+" == (30, "leftassoc")
getPrecAndAssoc "-" == (30, "leftassoc")

formatAnOpEntry1 :: String -> (String, String, String)

Parse an operator's definition to a triple ("relation" or "function" or "generic", precedence, associativity)

For example, ("+", "function:30:leftassoc") and ("ℙ", "generic::")

formatAnOpEntry1 "function:30:leftassoc"    == ("function", "30", "leftassoc")
formatAnOpEntry1 "generic::"                == ("generic", "", "")

isInfix :: Expression -> Bool

Check if an expression is an infix or not. Here we only consider InfixApp and ProdExpr as infix expressions, since other expressions have been parsed in *.y.

isExprInPars :: Expression -> Bool

Check if an expression is within a parenthesis, such as (a + b)

splitExpr :: Expression -> (Expression, String, Expression)

Get an infix expression's tuple (left expr, op, right expr)

splitExpr (ProdExpr ann [e1, e2])  == (e1, "x", e2)
splitExpr (FuncApplExpr ann11 (InfixApp ann12 (InInfixApp ann13 expr11 op1 expr12)))  == (expr11, op1, expr12)

replaceLeftExpr :: Expression -> Expression -> Expression

Replace an infix's left expression by the specified expression, and then return new expression

replaceRightExpr :: Expression -> Expression -> Expression

Replace an infix's right expression by the specified expression, and then return new expression

replaceInParent :: [Expression] -> Expression -> [Expression]

Replace all left expressions of a list of parents by the specified expression. Please note: (where parent shall be the left expression of grandparent)

For example,

replaceInParent [e1(e11 op1 e12), (e1 op2 e22)] e   == [e1'(e op1 e12), (e1' op2 e22)] 
replaceInParent [e1(e11 op1 e12), e2(e1 op2 e22), (e2 op3 e32)] e   == [e1'(e op1 e12), e2'(e1' op2 e22), (e2' op3 e32)] 

rightSub :: Expression -> [Expression] -> OpTableType (Either String (Expression, [Expression]))

The implementation of the RIGHT-SUBORDINATE algorithm

The brief description of the algorithm:

rightSub (if q is the parent of p)
     while (parents is not empty && the precedence of the q is larger than that of the op in current p)
         right subordination 
             p' = E1p Op (E2p Oq E2q) 
             remove q from parents, and then replace all q in its parents as p'

Reference: Handling Operator Precedence in Arithmetic Expression with Tree Transformations by WlLF R. LALONDE and JIM DES RlVlERES

returnExpr :: Expression -> [Expression] -> Expression

Return a final expression from current expression p and its parents

return e if parents is empty

return the last one in pars: it is the transformed one, if parents is not empty

returnExpr rootE []       == rootE
returnExpr e11 [(e11 op1 e12), ((e11 op1 e12) op2 e22), (((e11 op1 e12) op2 e22) op3 e32)] == (((e11 op1 e12) op2 e22) op3 e32)

rightTrans

Arguments

:: Expression
current expression to be processed

-> [Expression]
all parents that have been visited

-> OpTableType (Either String Expression)
return transformed expression

Right transformation function.

1. for the expression within a parenthesis, such as (e). Just transform this e, then return it

2. otherwise,

if it is an infix, transform its right expression (since we assume right expression is not involved in expression transformation according to precedence and associativities).

if it is the root expression, just shift

if it is not the root expression, call rightSub and then shift

if it is not an infix, transform e and then return transformed expression