Language.ISOZ.Parser.SynTransformerOne

Description

The first step of syntactical transformation to cope with some special Token or productions introduced only for temporary use.

It provides

synTraverse1Tree : traverse the original syntax tree

getAnnTerm : return a list of all annotations for a term

addAnnTerm : add an annotation to the list. If there's a same type, just override it

setAnnTerm : set the annotations of a term to the new list of annotations

Documentation

showPos :: [Ann] -> String

Show position information from a list of annotations

synTraverse1Tree :: SynTraverse1 a => a -> a

Traverse syntax tree.

getAnnTerm :: SynTraverse1 a => a -> [Ann]

Get a term's annotations.

getAnnTermAndRemoveExprInPars1 :: SynTraverse1 a => a -> [Ann]

Get a term's annotations but ExprInParAnn and ParserParDepthAnn are removed since it should not propagate to others.

getAnnTermAndRemoveExprInPars :: SynTraverse1 a => a -> [Ann]

Get a term's annotations but ExprInParAnn is removed since it should not propagate to others.

addAnnTerm :: SynTraverse1 a => Ann -> a -> a

Add an Ann to the term.

Please note: if there's a same type annotation, then just override it.

setAnnTerm :: SynTraverse1 a => [Ann] -> a -> a

Set the annotations of a term to specified annotations

class SynTraverse1 a where

Minimal complete definition

traverse1, getAnn, addAnn, setAnn

Methods

traverse1 :: a -> a

traverse1List :: [a] -> [a]

getAnn :: a -> [Ann]

addAnn :: Ann -> a -> a

setAnn :: [Ann] -> a -> a

addAnnList :: Ann -> [a] -> [a]

setAnnList :: [Ann] -> [a] -> [a]

isSchemaTextAnExpr :: SchemaText -> Maybe Expression