Language.ISOZ.Parser.SynTransformerTwo

Description

Syntactical transformation to cope with "Clause 12 Syntactic transformation rules" in ISO Z Standard.

It provides

synTraverse2Tree : traverse the syntax tree after synTraverse1Tree

Documentation

synTraverse2Tree :: SynTraverse2 a => a -> a

Traverse of the syntax tree for syntactical transformation

declToExpr :: Declaration -> Expression

(12.2.7.1) Transform a declaration to an expression

exprConsFromDecls :: [Declaration] -> Expression

(12.2.7.1 Declaration) Transform a list of declarations to a construct expression

declToNameExprPair :: [Declaration] -> [(NAME, Expression)]

Transform AbbrDecl in Declaration to a pair from NAME to Expression

transformSchema :: SchemaText -> Expression

(12.2.7 Schema text) transform a schema text to an expression

declNameToNAME :: DeclName -> NAME

Extrace NAME from a DeclName

declNamesToNAMEs :: [DeclName] -> [NAME]

Extrace NAMEs from DeclNames

refNameToNAME :: RefName -> NAME

Extrace NAME from a RefName

refNamesToNAMEs :: [RefName] -> [NAME]

Extrace NAMEs from RefNames

esssToNAME :: [ESSS] -> NAME

Extrace NAMEs from a list of ES and SS

eresreToNAME :: ERESRE -> NAME

Extrace NAME from a ERE or SRE

erepsrepToNAME :: EREPSREP -> NAME

Extrace NAME from a EREP or SREP

ersrToNAME :: ERSR -> NAME

Extrace NAME from a ER or SR

erpsrpToNAME :: ERPSRP -> NAME

Extrace NAME from a ERP or SRP

opNameToNAME :: OpName -> NAME

(12.2.8) Extrace NAME from a OpName

esssnToNAME :: [ESSS_N] -> NAME

Return (es1⋈...⋈esn)

esssnToNAME_i :: [ESSS_N] -> [NAME]

Return [i1,...,in]

eresrenToNAME :: ERESRE_N -> NAME

Extract NAME from a ERE_N or SRE_N

eresrenToNAME_i :: ERESRE_N -> NAME

ersrnToNAME :: ERSR_N -> NAME

ersrnToNAME_i :: ERSR_N -> NAME

genNameToNAME :: GenName -> (NAME, [NAME])

(12.2.9) Generic name

essseToNAME :: [ESSS_E] -> NAME

Return (es1⋈...esn)

transExpsList :: [Expression] -> Expression

(12.2.12) ExpressionList to {(1, e1), (2, e2), ...}

essseToNAME_e :: [ESSS_E] -> [Expression]

Return [e]

erepsrepeToNAME :: EREPSREP_E -> NAME

erepsrepeToNAME_e :: EREPSREP_E -> [Expression]

erpsrpeToNAME :: ERPSRP_E -> NAME

erpsrpeToNAME_e :: ERPSRP_E -> [Expression]

relationToPred :: Relation -> Predicate

(12.2.10) Relation operator application

All relation operator applications are transformed to annotated membership predicates.

eresreeToNAME :: ERESRE_E -> NAME

eresreeToNAME_e :: ERESRE_E -> [Expression]

ersreToNAME :: ERSR_E -> NAME

ersreToNAME_e :: ERSR_E -> [Expression]

removeStrokesFromOp :: String -> String

Remove strokes from an operator name

TODO: - 1. [x] op_1 - 2. [x] op^1 - 3. others ...

lookupOperatorMode :: String -> Map String Bool -> Maybe Bool

Lookup an operator's mode (function or generic) from the map

Notes: if (op) is an operator, then (op_1) should be an operator as well so we need to remove all strokes from the operator name, then lookup from the map

appToExpr :: Application -> Map String Bool -> Expression

2.11 Function and generic operator application

All function operator applications are transformed to annotated application expressions.

All generic operator applications are transformed to annotated generic instantiation expressions.

class SynTraverse2 a where

Minimal complete definition

traverse2

Methods

traverse2 :: a -> a

traverse2List :: [a] -> [a]