Language.ISOZ.Parser.SynTransformerZero

Description

Syntax transformation for ExprPredicate to Expression or Predicate depending on their context. In addition, check only schema expressions when they are expected. It provides

synTraverseZeroTree : traverse the original syntax tree

Documentation

synTraverseZeroTree :: SynTraverse0 a => a -> Either String a

Traverse a syntax tree for transformation of ExprPredicate.

class SynTraverse0 a where

Minimal complete definition

traverseZero

Methods

traverseZero :: a -> OpTableType (Either String a)

traverseZeroList :: [a] -> OpTableType (Either String [a])

ep2expr :: ExprPredicate -> OpTableType (Either String Expression)

Transformation from ExprPredicate to Expression.

ep2exprList :: [ExprPredicate] -> OpTableType (Either String [Expression])

Transformation from a list of ExprPredicates to a list of Expressions.

ep2exprSchema :: ExprPredicate -> OpTableType (Either String Expression)

Transformation from ExprPredicate to Expression and check if the transformed expression is a schema expression.

ep2pred :: ExprPredicate -> OpTableType (Either String Predicate)

Transformation from ExprPredicate to predicate.

ep2predList :: [ExprPredicate] -> OpTableType (Either String [Predicate])

Transform from a list of ExprPredicate to a list of Predicate.

isASchema :: String -> OpTableType Bool

Check if is a name is a schema.

andSchemaExprTrue :: Expression -> Expression -> OpTableType Bool

Check if both expressions are schema expressions.

isSchemaExpr :: Expression -> OpTableType Bool

Check if an expression is a schema expression.

notASchemaErrorExpr :: Expression -> String

Return an error message to indicate it is not a schema expression error.