Language.ISOZ.SecSolver

Description

SecSolver is a module having a lexer to parse and solve one input Z specification's all parent sections, including all standard toolkits. The order of paths to be checked for standard toolkits are

1. current path (the same folder with the input latex file)

2. ~/.ZParser/toolkits or env{HOME}/.ZParser/toolkits

3. env{ZPARSER_DIR}/toolkits

Documentation

secSolve

Arguments

:: String
path to input latex file

-> IO (Either String [String])
a list of pathes to its parent sections

get a list of an input latex file's all parent sections with absolute paths. For example, (secSolve "set_toolkit.tex") => ["/path/to/prelude.tex","/path/to/set_toolkit.tex"]