Language.ISOZ.Tex2Unicode

Description

Tex2Unicode is a module having a lexer to convert an input latex Z specification into corresponding unicode file according to Appendix A.2 of ISO Z Standard 2002.

Documentation

tex2Unicode

Arguments

:: String
path to input latex file

-> IO (Either String [Token])
a list of tokens

a lexer to convert a latex specification into a list of tokens

printResult :: Either String [Token] -> IO ()

print the lexed tokens

outputUnicode2File

Arguments

:: [Token]
a list of tokens

-> String
output file name

-> IO ()

output a list of unicode tokens into a file

tokensToUnicodeString :: [Token] -> String

extract Unicode strings from tokens and combine them into a line break separated string

tokensToUnicodeStringSpan

Arguments

:: [Token]
lexed tokens

-> Int
start line. In the beginning, it should start from line=1, column=1

-> Int
start column

-> (String, [(SrcSpan, SrcSpanInfo)])
returned pair from output unicode string to location information (map from output unicode string to original latex file)

extract Unicode strings from tokens and combine them into a line break separated string with location information mapping from target unicode string to source latex file

locToSourceSpan

Arguments

:: [(SrcSpan, SrcSpanInfo)]
location map from output unicode string to source latex file

-> Int
line number in unicode string

-> Int
column number in unicode string

-> Either String SrcSpanInfo
return source location information

search the location (line, column) from input location information, and return found source location of latex file