Language.ISOZ.TypeChecker.TypeInfer

Description

Annotate syntax sentence and typecheck by inference rules in Clause 13 "Type inference rules"

It provides

typeInferTraverseTree: traverse the syntax tree

Documentation

typeInferTraverseTree :: TypeInferTraverse a => a -> Either String a

traverse the syntax tree to type check it

if there is an error, then return Left error

otherwise, return a transformed and annotated syntax tree

class TypeInferTraverse a where

Traverse for typechecking

Minimal complete definition

traverseT, traverseTImplInst

Methods

traverseT :: a -> TypeC (Either String a)

traverseTList :: [a] -> TypeC (Either String [a])

traverseTImplInst :: a -> TypeC a

(13.2.3.3) Implicit Instantiation

traverseTImplInstList :: [a] -> TypeC [a]

preludeSectTypeEnv :: SectTypeEnv

The SectTypeEnv of the prelude section.

disjointNames :: [Paragraph] -> Either String Bool

Check if all names in the signatures of paragraphs are disjoint.

extractSign :: [Ann] -> Sig

Extract Signature from annotations.

hasAndExtractType :: [Ann] -> (Bool, Type)

Check if the annotation contains type annotation.

If so, just return it.

Otherwise, return false.

extractType :: [Ann] -> Type

Extract type from annotations.

getType2FromTerm :: SynTraverse1 a => a -> Type2

Get Type2 from a term's type annotation.

combineSectEnv :: [String] -> SectEnv -> Map String (String, Type)

Combine SectTypeEnv from several sections (as specified) into one SectTypeEnv (only map is return and no constructor)

combineSignsFromParas :: [Paragraph] -> Sig

Combine all signatures from paragraphs into one.

combineSigns :: [Sig] -> Sig

Combine a collection of signatures into one

signToSectTypeEnv :: String -> Sig -> Map String (String, Type)

Turn a section's signature into SectTypeEnv

withFunc :: (String, Type) -> (String, Type) -> (String, Type)

Used for Map.union to combine the same key (String, (String, Type)) (especially for SectTypeEnv).

For the same key, if they are from the same section with the same type, just return any one.

Otherwise, put a "Conflict" in front of concatenation of section names: s1 and s2 to indicate it is a conflict type or sections for a name.

Map.unionWith withFunc 
 (fromList [("name1", ("sect", T1)), ("name2", ("sect", T2))]) 
 (fromList [("name1", ("sect", T1)), ("name3", ("sect", T2)), ("name2", ("sect1", T2))]) 
== fromList [("name1", ("sect", T1)), ("name3", ("sect", T2)), ("name2", ("Conflictsectsect1", T2))] 

filterConflictFunc :: (String, Type) -> Bool

Check if it is a conflict name (with withFunc together)

assembleError :: String -> [Ann] -> String

Assemble error messages

returnDuplicates :: Eq a => [a] -> [a]

Return duplicates elements in the list. Even one element occurs several times in the list, it still just returns once.

For example,

returnDuplicates [a, b, a, c, c, c] = [a, c]

extractDecl :: SchemaText -> [Declaration]

Extract declarations from a schema text.

genTypeInst

Arguments

:: Type
Generic type

-> [Type2]
A square-bracketed list of argument types

-> Type2
Instantiated type

2.3.1 Generic type instantiation

carrier :: Type2 -> Expression

2.3.2 Carrier set

isCompatible :: (Eq a, Ord k) => Map k a -> Map k a -> Bool

Check whether two signatures are incompatible or not.

Compatible relations [e1 ≈ e2] is equal to [(dom e2)▹e1 = (dom e1)▹e2].

listDiff :: (Eq t, Foldable t1) => [t] -> t1 t -> [t]

Return the difference of two lists, like the set minus (l1 - l2).

Finally it is equal to a modified list one in which all elements, that are in the list two as well, are removed.

listDiff [2, 1, 3, 4] [3, 5, 7, 1] == [2, 4]

dropList :: (Eq a1, Foldable t) => Map a1 a -> t a1 -> Map a1 a

Drop all entries in which the key is in the list ("names").

unify :: Type2 -> Type2 -> TypeC Bool

Unify two types.

equalListTest :: Eq a => [a] -> [a] -> Bool

Check if two lists have the same elements which might be not in the same order.

equalListTestStr :: [String] -> [String] -> [(String, String)] -> (Bool, [(String, String)])

updateTypeVarInTypeAnn :: [Ann] -> TypeC [Ann]

Replace all type variables in annotations by its instantiated type